fix: Arrumando os problemas do merge com a develop
