+++ src/components/escolasCadastradas/tabela/Filtro/index.tsx
