+++ src/components/escolasCadastradas/tabela/Tabela/styles.css
