+++ src/pages/cadastro/CadastrarEscola/index.tsx
@@ -1,9 +1,23 @@
+import React, { useContext, useEffect } from "react";
 import Header from "../../../components/Header";
+import Footer from "../../../components/Footer";
 import RegS from "../../../components/cadastrarEscolas/CadastrarEscola";
 import { FiltroProvider } from "../../../context/FiltroTabela";
 import "../../../styles/App.css";
+import { AuthContext } from "../../../provider/Autenticacao";
+import { useNavigate } from "react-router-dom";
+import { Permissao } from "../../../models/auth";
 
 function CadastrarEscola() {
+  const navigate = useNavigate();
+  const { temPermissao } = useContext(AuthContext);
+
+  useEffect(() => {
+    if (!temPermissao(Permissao.EscolaCadastrar)) {
+      navigate("/");
+    }
+  }, []);
+
   return (
     <div className="App">
       <Header />
@@ -14,6 +28,7 @@ function CadastrarEscola() {
           </FiltroProvider>
         </div>
       </div>
+      <Footer />
     </div>
   );
 }
