+++ src/tests/components/Table.spec.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable testing-library/render-result-naming-convention */
 import { render, fireEvent } from "@testing-library/react";
-import CustomTable, { CustomTableRow } from "../../components/Table/Table";
+import CustomTable, { CustomTableRow } from "../../components/Table";
 
 describe("CustomTable Component", () => {
   const rowsData = [
